+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardResult.kt
@@ -7,6 +7,9 @@ import android.net.Uri
  *
  * @property [orderId] associated order ID.
  * @property [liabilityShift] Liability shift value returned from 3DS verification
+ * @property [status] status of the order
+ * @property [didAttemptThreeDSecureAuthentication] 3DS verification was attempted.
+ * Use v2/checkout/orders/{orderId} in your server to get verification results.
  */
 data class CardResult(
     val orderId: String,
@@ -14,6 +17,13 @@ data class CardResult(
     /**
      * @suppress
      */
+    @Deprecated("Use status instead.")
     val deepLinkUrl: Uri? = null,
-    val liabilityShift: String? = null
+
+    @Deprecated("Use didAttemptThreeDSecureAuthentication instead.")
+    val liabilityShift: String? = null,
+
+    val status: String? = null,
+
+    val didAttemptThreeDSecureAuthentication: Boolean = false
 )